pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["MaxKushSolo <maxekushnir@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.114.0"
pydantic = {extras = ["email"], version = "^2.9.1"}
SQLAlchemy = "^2.0.34"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
uvicorn = "^0.30.6"
sqlalchemy-utils = "^0.41.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
